Rename FileButton to FileWithDownload
